requirements.txt
-r requirements-encode.txt
Django==4.2.20
Pillow==10.3.0
django-tinymce==4.1.0
django-tagulous==2.1.0
django-modeltranslation==0.19.11
django-cas-ng==5.0.1
ldap3==2.9.1
django-simple-captcha==0.6.0
urllib3==2.2.3
elasticsearch==8.17.2
djangorestframework==3.15.2
django-filter==24.3
markdown==3.4.1
sorl-thumbnail==12.11.0
django-lti-provider==1.0.0
django-auth-mixins==1.0.1
pandas
django-shibboleth-remoteuser
django-chunked-upload==2.0.0
django-select2
django-redis
chardet==5.2.0
django-static-compress>=2.0.0
mozilla_django_oidc==4.0.1
ics==0.7.2
django-honeypot==1.2.1
qrcode
ralph-malph==5.0.1
pydantic==2.10.2
bleach==6.2.0
pytubefix==8.12.0
django-redis-sessions
paramiko~=3.5.0
djangorestframework-simplejwt==5.3.1
django-pwa==2.0.1
django-webpush==0.3.6
defusedxml==0.7.1
